Trying some parallel stuff
git-svn-id: https://geuz.org/svn/gmsh/trunk@16039 59c8cabf-10d6-4a28-b43f-3936b85956c4

+++ projects/small_fem/postprocessing/WriterMshInterpolation.cpp
@@ -5,7 +5,7 @@ using namespace std;
 void WriterMsh::writeInterpolationScheme(void) const{
   // Some Temp Value
   const fullMatrix<double>& coef = lBasis->getCoefficient();
-  const fullMatrix<int>&    mono = lBasis->getMonomial();
+  const fullMatrix<double>& mono = lBasis->getMonomial();
 
   const unsigned int nRowCoef = coef.size1();
   const unsigned int nColCoef = coef.size2();
